After crit + close all other non-gallery divs
This code has been altered to remove the non-gallery divs when the
gallery expands. This greatly improves performance, especially on the
sluggish Safari browser.

Now the trick will be to figure out the javascript to scroll to gallery
anchor, then expand the gallery height, then close other divs and
expand the gallery modal.